Fix get_headers function (don't reinitialise string tokeniser)
The get_headers function was reinitialising the string tokeniser before
finishing with the original string. This caused it to fail to retrieve a
Location: header if the Set-Cookie came before it, and thus re-request the
original URL  This resulted in nonce_used OAuth errors appearing in the
ESS logs every time a user accessed Echo via the EchoLink module.